This commit aims to: 1.) Update some styling 2.) "Pull up" the pie chart from it's own component into the parent Admin component for better refreshing of the chart during drag and drop scenarios 3.) Adds in live charting for on break staff 4.) General code cleaning and updating.

+++ YoumaconSecurityOps.Web.Client/Models/PieChartDataModel.cs
@@ -3,6 +3,6 @@ public class PieChartDataModel
 {
     public String Location { get; set; }
     
-    public Int32 Shifts { get; set; }
+    public Double Shifts { get; set; }
 }
